Added some small changes and tweaks.
We now wait on game over using pygame.time.delay() so the computer knows the application is still responding. You can no longer press 'p' on the menu screen to start playing. You can now automatically select the correct answer by pressing 'c' in game. And last but not least, the horse now starts at a speed of 5 and accelerates towards a speed of 10 with each correct answer.